requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
rank-bm25
tqdm
python-dotenv
